Split ControllerStructs into a Controllers folder; move WiiDrums/WiiGuitar into Controllers folder
