+++ Nintroller/Controllers/WiiDrums.cs
