+++ Nintroller/Controllers/WiiGuitar.cs
