+++ Nintroller/Nintroller.csproj
@@ -58,18 +58,24 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Controllers/BalanceBoard.cs" />
+    <Compile Include="Controllers/ClassicController.cs" />
+    <Compile Include="Controllers/ClassicControllerPro.cs" />
+    <Compile Include="Controllers/Nunchuk.cs" />
+    <Compile Include="Controllers/ProController.cs" />
+    <Compile Include="Controllers/WiiDrums.cs" />
+    <Compile Include="Controllers/WiiGuitar.cs" />
+    <Compile Include="Controllers/Wiimote.cs" />
+    <Compile Include="Controllers/WiimotePlus.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="DeviceManagement.cs" />
     <Compile Include="Calibrations.cs" />
     <Compile Include="ComponentStructs.cs" />
-    <Compile Include="ControllerStructs.cs" />
     <Compile Include="Enumerators.cs" />
     <Compile Include="INPUT_NAMES.cs" />
     <Compile Include="Interfaces.cs" />
     <Compile Include="Nintroller.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WiiDrums.cs" />
-    <Compile Include="WiiGuitar.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
